📝 Prettify init instance guide (#331)
* :memo: Prettify init instance

* :lipstick: Prettify logging

* :green_heart: Fix

* :green_heart: Fix

* :green_heart: Fix

* :green_heart: Fix

* :green_heart: Fix

* :green_heart: Fix

* :green_heart: Fix

* :heavy_plus_sign: Add lnschema_bionty

* :arrow_down: Do not upgrade jupyter_client

* :arrow_down: Downgrade

* :memo: Downgrade

* :construction_worker: Double pin

* add ipykernel

* :arrow_up: Fix nbproject_test

* :heavy_plus_sign: Add lnschema-wetlab

* :green_heart: Fix

---------

+++ lndb/_delete.py
@@ -12,6 +12,11 @@ def delete(instance_name: str):
     """Delete an instance."""
     logger.info(f"Deleting instance {settings.user.handle}/{instance_name}")
     settings_file = instance_settings_file(instance_name, settings.user.handle)
+    if not settings_file.exists():
+        raise RuntimeError(
+            "Instance settings do not exist locally. Did you provide a wrong instance"
+            " name? Could you try loading it?"
+        )
     isettings = load_instance_settings(settings_file)
 
     delete_settings(settings_file)
@@ -19,12 +24,14 @@ def delete(instance_name: str):
     logger.info(
         f"    consider deleting your stored data manually: {isettings.storage.root}"
     )
+    if isettings.dialect == "sqlite":
+        if isettings._sqlite_file.exists():
+            isettings._sqlite_file.unlink()
+            logger.info("    deleted '.lndb' sqlite file")
+        else:
+            logger.info("    '.lndb' sqlite file does not exist")
     if isettings.is_remote:
         logger.info("    please manually delete your remote instance on lamin.ai")
-    else:
-        if isettings.dialect == "sqlite":
-            isettings._sqlite_file.unlink()
-            logger.info("    deleted sqlite file")
 
 
 def delete_cache(cache_dir: Path):
@@ -34,5 +41,8 @@ def delete_cache(cache_dir: Path):
 
 
 def delete_settings(settings_file: Path):
-    settings_file.unlink()
-    logger.info("    instance settings deleted")
+    if settings_file.exists():
+        settings_file.unlink()
+        logger.info("    instance settings '.env' deleted")
+    else:
+        logger.info("    instance settings '.env' do not exist locally")
